+++ temporal_normalization/commons/temporal_models.py
@@ -1,6 +1,6 @@
 import json
 
-from py4j.java_gateway import JavaObject
+from py4j.java_gateway import JavaObject, JavaGateway
 
 from temporal_normalization.commons.temporal_types import TemporalType
 
@@ -14,6 +14,7 @@ class TemporalExpression:
         is_valid (bool): A flag that specifies whether the text processed
             through timespan-normalization library is a temporal expression.
         input_value (str or None): The original temporal expression before processing.
+        prepared_value (str or None): The temporal expression after processing.
         time_series (list[TimeSeries]): The list of normalized temporal expressions.
         matches (list[str]): A unique list of matched values found in the normalized
             entities.
@@ -23,10 +24,12 @@ class TemporalExpression:
         serialize = java_object.serialize()
         json_obj = json.loads(serialize)
 
+        # fmt: off
         self.is_valid = TemporalExpression.is_valid_json(json_obj)
         self.input_value: str | None = json_obj["inputValue"] if self.is_valid else None
+        self.prepared_value: str | None = json_obj["preparedValue"] if self.is_valid else None
         self.time_series: list[TimeSeries] = [
-            TimeSeries(item) for item in json_obj["timeSeries"]
+            TimeSeries(item, self.input_value, self.prepared_value) for item in json_obj["timeSeries"]
         ] if self.is_valid else []
         self.matches: list[str] = list(
             set(
@@ -37,6 +40,7 @@ class TemporalExpression:
                 ]
             )
         )
+        # fmt: on
 
     def __str__(self):
         if self.input_value is None:
@@ -52,12 +56,40 @@ class TemporalExpression:
         return "inputValue" in json_obj and "timeSeries" in json_obj
 
 
+def extract_temporal_expressions(
+    gateway: JavaGateway, text: str
+) -> list[TemporalExpression]:
+    """
+    Extracts valid temporal expressions from the given text using the Java temporal
+    normalization gateway.
+
+    Args:
+        gateway (JavaGateway): Active Py4J gateway connected to the Java temporal
+            normalization process.
+        text (str): Input text from which to extract temporal expressions.
+
+    Returns:
+        list[TemporalExpression]: A list containing valid temporal expressions.
+    """
+
+    expressions: list[TemporalExpression] = []
+    java_object = gateway.jvm.ro.webdata.normalization.timespan.ro.TimeExpression(text)
+    temporal_expression = TemporalExpression(java_object)
+
+    if temporal_expression.is_valid:
+        expressions.append(temporal_expression)
+
+    return expressions
+
+
 class TimeSeries:
     """
     A data structure representing a temporal expression that has been normalized
     into a list of periods and temporal edges.
 
     Attributes:
+        input_value (str or None): The original temporal expression before processing.
+        prepared_value (str or None): The temporal expression after processing.
         edges (list[EdgeModel]): A list of temporal intervals represented as edges.
         periods (list[DBpediaModel]): A list of normalized DBpedia entities
             extracted from the expression.
@@ -65,7 +97,9 @@ class TimeSeries:
             entities.
     """
 
-    def __init__(self, data: dict):
+    def __init__(self, data: dict, input_value: str, prepared_value: str):
+        self.input_value = input_value
+        self.prepared_value = prepared_value
         self.edges: EdgeModel = EdgeModel(data["edges"]) if "edges" in data else None
         self.periods: list[DBpediaModel] = (
             [DBpediaModel(item) for item in data["periods"]]
@@ -80,10 +114,12 @@ class TimeSeries:
         return f"TimeSeries(edges={self.edges}, periods={self.periods})"
 
     def serialize(self, indent: str = ""):
+        # fmt: off
         return (
             f"{indent}Edges: {self.edges}\n"
             f"{indent}Periods: {self.periods}"
         )
+        # fmt: on
 
 
 class DBpediaModel: