+++ temporal_normalization/index.py
@@ -1,17 +1,26 @@
 import re
+import subprocess
+from pathlib import Path
 
+from py4j.java_gateway import JavaGateway
 from spacy import Language
 from spacy.tokens import Doc, Span
+from spacy.tokens._retokenize import Retokenizer
 from spacy.util import filter_spans
 
 from temporal_normalization import TimeSeries
-from temporal_normalization.commons.temporal_models import TemporalExpression
-from temporal_normalization.process.java_process import start_process
+from temporal_normalization.commons.temporal_models import (
+    extract_temporal_expressions,
+    TemporalExpression,
+)
+from temporal_normalization.process.java_process import start_conn, close_conn
 
 try:
+
     @Language.factory("temporal_normalization")
     def create_normalized_component(nlp, name):
         return TemporalNormalization(nlp, name)
+
 except AttributeError:
     # spaCy 2.x
     pass
@@ -21,7 +30,7 @@ class TemporalNormalization:
     """
     spaCy pipeline component for identifying and annotating temporal expressions in text.
 
-    This component calls the ``start_process`` method to extract temporal expressions, then
+    This component calls the ``start_conn`` method to extract temporal expressions, then
     aligns the matches with spaCy tokens using retokenization and sets a custom attribute
     containing associated time series metadata.
     """
@@ -40,6 +49,11 @@ class TemporalNormalization:
         Span.set_extension(TemporalNormalization.__FIELD, default=None, force=True)
         self.nlp = nlp
 
+        root_path = str(Path(__file__).resolve().parent.parent)
+        java_process, gateway = start_conn(root_path)
+        self.java_process: subprocess.Popen = java_process
+        self.gateway: JavaGateway = gateway
+
     def __call__(self, doc: Doc) -> Doc:
         """
         Apply the component to a spaCy Doc object.
@@ -54,13 +68,16 @@ class TemporalNormalization:
             Doc: The modified Doc object with temporal expressions processed.
         """
 
-        expressions: list[TemporalExpression] = []
-        start_process(doc.text, expressions)
+        expressions: list[TemporalExpression] = extract_temporal_expressions(
+            self.gateway, doc.text
+        )
         str_matches: list[str] = _prepare_str_patterns(expressions)
-
         _retokenize(doc, str_matches, expressions)
 
         return doc
+
+    def __del__(self):
+        close_conn(self.java_process, self.gateway)
 
 
 def _prepare_str_patterns(expressions: list[TemporalExpression]) -> list[str]:
@@ -101,7 +118,8 @@ def _retokenize(
                                                 each with time series metadata.
     """
 
-    regex_matches: list[str] = [rf"{item}" for item in str_matches]
+    # TODO: WIP
+    regex_matches: list[str] = [rf"{re.escape(item)}" for item in str_matches]
     pattern = f"({'|'.join(regex_matches)})"
     matches = (
         list(re.finditer(pattern, doc.text, re.IGNORECASE))
@@ -126,6 +144,7 @@ def _retokenize(
                 if token.idx + len(token.text) == end_char:
                     end_token = token.i
 
+            # fmt: off
             if start_token is not None and end_token is not None:
                 # use exact token boundaries to create a custom `Span` for well-defined
                 # time expressions with known character offsets.
@@ -144,6 +163,8 @@ def _retokenize(
                         matched_ts = [ts for ts in time_series if _is_substring(entity.text, ts.matches)]
                         _retokenize_entity(doc, matched_ts, entity, True, retokenized_entities, retokenizer)
 
+            # fmt: on
+
 
 def _retokenize_entity(
     doc: Doc,
@@ -151,7 +172,7 @@ def _retokenize_entity(
     entity: Span,
     existed_entity: bool,
     retokenized_entities: list[Span],
-    retokenizer: Doc.retokenize,
+    retokenizer: Retokenizer,
 ) -> None:
     """
     Retokenizes and enriches a temporal entity span with matched time series data.
@@ -172,6 +193,8 @@ def _retokenize_entity(
     _assign_time_series(matched_ts, entity, existed_entity)
     _update_doc_ents(doc, entity)
     _merge_entity(doc, entity, retokenized_entities, retokenizer)
+
+    return None
 
 
 def _assign_time_series(
@@ -214,7 +237,7 @@ def _merge_entity(
     doc: Doc,
     entity: Span,
     retokenized_entities: list[Span],
-    retokenizer: Doc.retokenize,
+    retokenizer: Retokenizer,
 ) -> None:
     """
     Merges a custom entity span into the spaCy Doc if it is not already part of
@@ -223,7 +246,7 @@ def _merge_entity(
     Args:
         entity (Span): The named entity to enrich.
         retokenized_entities (list): Accumulator for entities that require retokenization.
-        retokenizer (Doc.retokenize): The spaCy retokenizer context.
+        retokenizer (Retokenizer): The spaCy retokenizer context.
     """
 
     if entity not in doc.ents: