+++ validation/index.py
@@ -1,5 +1,0 @@
-from validation import validate_dataset
-
-if __name__ == "__main__":
-    validate_dataset("validation")
-    validate_dataset("test")